backend/requirements.txt
fastapi
uvicorn
anthropic
python-dotenv
pydantic
openai==1.12.0